Add initial Deck class
The Deck class is used to initialize a new (52 card) deck. There are
private methods that define the suits and ranks of cards so they can
easily be modified later on.

This commit also defines the `find_by_suit_or_rank` method, which should
be used when scoring decks later on.

Initial for the Deck class are also included in this commit.