Add SearchUtilities module
It should be trivial to search for a attributes of either a hand or
a deck. The SearchUtilities module will be used to house methods that
can be used by any other classes I need to make in the PokerHands
module.

This also moves the `find_by_suit_or_rank` method to the SearchUtilities
module, because I can already tell I'll want to use in in the Hand
class.